refactor(button): move button outside of components folder
